# docs/api-documentation.yaml
openapi: 3.0.0
info:
  title: Training Management System API
  description: Complete REST API for managing corporate training programs with trainers, projects, engagements, and attendance tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@trainermanagement.com
servers:
  - url: http://localhost:8000/v1
    description: Local development server
  - url: https://api.trainermanagement.com/v1
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        client_name:
          type: string
        primary_contact_name:
          type: string
        primary_contact_email:
          type: string
        primary_contact_phone:
          type: string
        project_type:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        description:
          type: string
        status:
          type: string
          enum: [active, upcoming, completed]
        engagements:
          type: array
          items:
            $ref: '#/components/schemas/Engagement'
    Engagement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        domain:
          type: string
        training_type:
          type: string
        total_students:
          type: integer
        students_per_batch:
          type: integer
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [active, upcoming, completed]
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
    Batch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        engagement_id:
          type: string
          format: uuid
        batch_number:
          type: integer
        students:
          type: integer
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [confirmed, pending, awaiting_confirmation]
        trainer_id:
          type: string
          format: uuid
        confirmed_at:
          type: string
          format: date-time
        confirmed_by:
          type: string
          format: uuid
...
security:
  - BearerAuth: []
paths:
...
  /projects/:
    get:
      tags:
        - Projects
      summary: Get all projects
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

    post:
      tags:
        - Projects
      summary: Create new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                client_name:
                  type: string
                primary_contact:
                  type: object
                  properties:
                    name:
                      type: string
                    email:
                      type: string
                    phone:
                      type: string
                project_type:
                  type: array
                  items:
                    type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                description:
                  type: string
      responses:
        '201':
          description: Project created
  /projects/{project_id}:
    get:
      tags:
        - Projects
      summary: Get project by ID
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details with engagements and batches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{project_id}/engagements:
    post:
      tags:
        - Engagements
      summary: Create engagement for project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                domain:
                  type: string
                training_type:
                  type: string
                total_students:
                  type: integer
                students_per_batch:
                  type: integer
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Engagement created with auto-generated batches
...